% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repr.R
\name{repr_*.gvis}
\alias{repr_*.gvis}
\alias{repr_html.gvis}
\title{repr_html.gvis}
\usage{
repr_html.gvis(obj, ...)
}
\arguments{
\item{obj}{A gvis object.}

\item{...}{unused}
}
\value{
Returns the html representation of the gvis object as string
}
\description{
This function lets you use googleVis charts as html output in e.g. a Jupyter Notebook.
}
\details{
Output a gvis chart as an embedded html chart in the Jupyter ecosystem
}
\author{
Jan Schulz, \email{jasc@gmx.net}
}
\keyword{jupyter}

