\name{Exports}
\alias{Exports}
\docType{data}
\title{
Exports: googleVis example data set
}
\description{
  Example data set to illustrate the use of the googleVis package.
}
\usage{data(Exports)}
\format{
  A data frame with 10 observations on the following 3 variables.
  \describe{
    \item{\code{Country}}{a factor with levels \code{Brazil}, \code{Germany} ...}
    \item{\code{Profit}}{a numeric vector}
    \item{\code{Online}}{a logical vector}
  }
}
\examples{
data(Exports)
str(Exports)
}
\keyword{datasets}
