\name{googleVis-package}
\alias{googleVis-package}
\alias{googleVis}
\docType{package}
\title{
Using the Google Visualisation API with R
}
\description{
googleVis is an R package providing an interface between R and the Google Visualisation API. 
The functions of the package allow the user to visualise data stored in a R data.frames with the Google Visualisation API.

The output is html code, which when run on a web server, generates a dynamic flash chart.
  
} 
\details{
\tabular{ll}{
Package: \tab googleVis\cr
Type: \tab Package\cr
Version: \tab 0.2.0\cr
Date: \tab 20010-10-10\cr
License: \tab GPL version 2 or later\cr
}
}
\author{
Markus Gesmann, Diego de Castillo

Maintainer: Markus Gesmann <markus.gesmann@gmail.com>
}
\references{
\itemize{
\item Google Visualisation API \url{http://code.google.com/apis/visualization/interactive_charts.html}
\item Google Terms of Use \url{http://code.google.com/apis/visualization/documentation/gallery/treemap.html}
\item googleVis \url{http://code.google.com/p/google-motion-charts-with-r/}
}
}

\examples{
\dontrun{
  demo(googleVis)
  }
}
\keyword{ package }


