% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gvisWordTree.R
\name{gvisWordTree}
\alias{gvisWordTree}
\title{Google Word Tree with R}
\usage{
gvisWordTree(data, textvar = "", sizevar = "", stylevar = "",
  options = list(), chartid)
}
\description{
The gvisWordTree function reads a data.frame and creates text outpute referring to the Google Visualisation API,
which can be included into a web page, or as a stand-alone page. The actual chart is rendered by the web browser.
}
